+++ CHANGELOG.md
@@ -27,6 +27,7 @@ straightforward as possible.
 - Half-uploaded files bug
 - force video loading in chat
 - extra invites for checkpoints into cargo room
+- cargo camera sensor input stability
 
 ## 2023-06-30_ad580b9___2023-06-29_6ab0603